move geoidheight cloud function out of index
